added tabs and updated tests, fixes #16

+++ apps/backend/controllers/DestinationController.php
@@ -49,6 +49,12 @@ class DestinationController extends \Phalcon\Mvc\Controller
             $destinationTabs = array();
             foreach ($this->request->getPost('tabs') as $tabType => $tabDescription)
             {
+                // not added, pass
+                if (!$tabDescription)
+                {
+                    continue;
+                }
+                
                 $destinationTab = new \Robinson\Backend\Models\Tabs\Destination();
                 $destinationTab->setType($tabType)
                     ->setTitle($destinationTab->resolveTypeToTitle())
@@ -77,6 +83,8 @@ class DestinationController extends \Phalcon\Mvc\Controller
         ));
         
         $this->view->setVar('categories', $categories);
+        
+        $this->view->tabs = $this->getDI()->getShared('config')->application->destination->tabs->toArray();
     }
     
     /**
@@ -101,6 +109,7 @@ class DestinationController extends \Phalcon\Mvc\Controller
             
             foreach ($this->request->getPost('tabs') as $tabType => $tabDescription)
             {
+                $tabDescription = trim($tabDescription);
                 $tab = $destination->getTabs(array
                 (
                     'type = :type:',
@@ -110,9 +119,24 @@ class DestinationController extends \Phalcon\Mvc\Controller
                     ),
                 ))->getFirst();
                 
-                if (!$tabDescription)
+                // new tab
+                if (!$tab && $tabDescription)
+                {
+                   $tab = new \Robinson\Backend\Models\Tabs\Destination();
+                   $tab->setType($tabType)
+                       ->setTitle($tab->resolveTypeToTitle());
+                }
+       
+                // deleted tab
+                if ($tab && !$tabDescription)
                 {
                     $tab->delete();
+                    continue;
+                }
+              
+                // never existed and wasnt entered
+                if (!$tab && !$tabDescription)
+                {
                     continue;
                 }
                 
@@ -160,6 +184,13 @@ class DestinationController extends \Phalcon\Mvc\Controller
         
         $this->view->categories = $categories;
         $this->view->destination = $destination;
+        $this->view->tabs = $this->getDI()->getShared('config')->application->destination->tabs->toArray();
+        
+        $tabs = $destination->getTabs();
+        foreach ($tabs as $tab)
+        {
+            $this->tag->setDefault('tabs[' . $tab->getType() . ']', $tab->getDescription());
+        }
         $this->tag->setDefault('categoryId', $destination->getCategoryId());
         $this->tag->setDefault('destination', $destination->getDestination());
         $this->tag->setDefault('description', $destination->getDescription());
